+++ lib/djv/App/SettingsTool.cpp
@@ -104,17 +104,17 @@ namespace djv
             p.bellows["FFmpeg"]->setWidget(vLayout2);
 #endif // TLRENDER_FFMPEG_PLUGIN
 
-#if defined(TLRENDER_FFMPEG_PIPE)
-            p.bellows["FFmpegPipe"] = ftk::Bellows::create(context, "FFmpeg Pipe", vLayout);
+#if defined(TLRENDER_FFMPEG_CMD)
+            p.bellows["FFmpegCmd"] = ftk::Bellows::create(context, "FFmpeg Command", vLayout);
             vLayout2 = ftk::VerticalLayout::create(context, vLayout);
             vLayout2->setMarginRole(ftk::SizeRole::Margin);
             ftk::Label::create(
                 context,
                 "Changes are applied to new files.",
                 vLayout2);
-            ui::FFmpegPipeSettingsWidget::create(context, settingsModel, vLayout2);
-            p.bellows["FFmpegPipe"]->setWidget(vLayout2);
-#endif // TLRENDER_FFMPEG_PIPE
+            ui::FFmpegCmdSettingsWidget::create(context, settingsModel, vLayout2);
+            p.bellows["FFmpegCmd"]->setWidget(vLayout2);
+#endif // TLRENDER_FFMPEG_CMD
 
 #if defined(TLRENDER_USD)
             p.bellows["USD"] = ftk::Bellows::create(context, "USD", vLayout);